Move strip_extras() to utils
