+++ src/pip/_internal/req/constructors.py
@@ -38,7 +38,7 @@ from pip._internal.utils.pylock import (
     package_vcs_requirement_url,
     package_wheel_requirement_url,
 )
-from pip._internal.utils.urls import path_to_url
+from pip._internal.utils.urls import path_to_url, strip_extras
 from pip._internal.vcs import is_url, vcs
 
 __all__ = [
@@ -52,24 +52,6 @@ logger = logging.getLogger(__name__)
 # All standard version specifier operators
 # https://packaging.python.org/en/latest/specifications/version-specifiers/#id5
 operators = ("~=", "==", "!=", "<=", ">=", "<", ">", "===")
-
-
-def strip_extras(path: str) -> tuple[str, set[str]]:
-    m = re.match(r"^(.+)(\[[^\]]+\])$", path)
-    extras = None
-    if m:
-        path_no_extras = m.group(1).rstrip()
-        extras = m.group(2)
-    else:
-        path_no_extras = path
-
-    return path_no_extras, _convert_extras(extras)
-
-
-def _convert_extras(extras: str | None) -> set[str]:
-    if not extras:
-        return set()
-    return get_requirement("placeholder" + extras.lower()).extras
 
 
 def _set_requirement_extras(req: Requirement, new_extras: set[str]) -> Requirement:
